Logs in Cron - Jobs folder is updated , they are now industry standard, error handlers are used to increase reusablitiy of the code

+++ cron-jobs/delete-deactivated-users.job.js
@@ -2,21 +2,30 @@ const cron = require("node-cron");
 const UserModel = require("../models/user.model");
 const { logWithTime } = require("../utils/time-stamps.utils");
 const { userCleanup } = require("../configs/cron.config");
+const { errorMessage } = require("../configs/error-handler.configs");
 
 const deleteDeactivatedUsers = async () => {
   try {
     if(!userCleanup.enable)return;
+    if (!userCleanup.deactivatedRetentionDays || userCleanup.deactivatedRetentionDays < 1) {
+      logWithTime("⚠️ Invalid retention days configuration. Skipping user cleanup.");
+      return;
+    }
     const cutoffDate = new Date(Date.now() - userCleanup.deactivatedRetentionDays * 24 * 60 * 60 * 1000);
+    logWithTime("📅 [CRON-JOB] ➤ Deativated Users Cleanup Started...");
     const result = await UserModel.deleteMany({
       isActive: false,
       lastDeactivatedAt: { $lt: cutoffDate },
       userType: "CUSTOMER"
     });
-
-    logWithTime(`🗑️ Account Deletion Job: ${result.deletedCount} users hard deleted (inactive > ${userCleanup.deactivatedRetentionDays} days).`);
+    if(result.deletedCount === 0){
+      logWithTime(`📭 No users eligible for deletion (deactivated more than ${userCleanup.deactivatedRetentionDays} days).`);
+    }else {
+      logWithTime(`🗑️ Account Deletion Job: ${result.deletedCount} users hard deleted (inactive > ${userCleanup.deactivatedRetentionDays} days).`);
+    }
   } catch (err) {
-    logWithTime("❌ Error in deleting old deactivated users.");
-    console.error(err);
+    logWithTime("❌ Internal Error in deleting old deactivated users.");
+    errorMessage(err);
   }
 };
